+++ packages/schema-command/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @collidor/command
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @collidor/command@6.0.2
+
 ## 0.1.0
 
 ### Minor Changes